TGP-2816: moved views
