+++ app/controllers/profile/ProfileController.scala
@@ -28,7 +28,7 @@ import repositories.SessionRepository
 import utils.SessionData.{dataAdded, dataRemoved, dataUpdated, pageUpdated}
 import viewmodels.checkAnswers._
 import viewmodels.govuk.summarylist._
-import views.html.ProfileView
+import views.html.profile.ProfileView
 
 import javax.inject.Inject
 import scala.concurrent.{ExecutionContext, Future}
