+++ app/controllers/profile/UkimsNumberChangeController.scala
@@ -20,7 +20,7 @@ import controllers.actions._
 import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendBaseController
-import views.html.UkimsNumberChangeView
+import views.html.profile.UkimsNumberChangeView
 
 import javax.inject.Inject
