+++ app/views/profile/CyaCreateProfileView.scala.html
@@ -45,7 +45,7 @@
     </div>
 
     @govukSummaryList(list)
-    @formHelper(action = profile.routes.CyaCreateProfileController.onSubmit()) {
+    @formHelper(action = controllers.profile.routes.CyaCreateProfileController.onSubmit()) {
         @govukButton(ButtonViewModel(messages("site.save.continue")))
     }
 }
