+++ app/views/profile/CyaNewUkimsNumberView.scala.html
@@ -35,7 +35,7 @@
 
     @govukSummaryList(list)
 
-    @formHelper(action = profile.routes.CyaNewUkimsNumberController.onSubmit()) {
+    @formHelper(action = controllers.profile.routes.CyaNewUkimsNumberController.onSubmit()) {
         @govukButton(ButtonViewModel(messages("site.save.continue")))
     }
 }
