+++ app/views/profile/UkimsNumberView.scala.html
@@ -61,7 +61,7 @@
 
     @if(isCreateJourney) {
         <p class="govuk-body">
-            <a href="@profile.routes.UkimsKickOutController.onPageLoad().url" class="govuk-link">@messages("ukimsNumber.linkText")</a>
+            <a href="@controllers.profile.routes.UkimsKickOutController.onPageLoad().url" class="govuk-link">@messages("ukimsNumber.linkText")</a>
         </p>
     }
