+++ test/controllers/profile/CreateProfileSuccessControllerSpec.scala
@@ -14,27 +14,38 @@
  * limitations under the License.
  */
 
-package controllers
+package controllers.profile
 
 import base.SpecBase
+import connectors.TraderProfileConnector
+import org.mockito.ArgumentMatchers.any
+import org.mockito.Mockito.when
+import org.scalatestplus.mockito.MockitoSugar.mock
+import play.api.inject
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
-import views.html.UkimsNumberChangeView
 
-class UkimsNumberChangeControllerSpec extends SpecBase {
+import scala.concurrent.Future
 
-  "UkimsNumberChangeController Controller" - {
+class CreateProfileSuccessControllerSpec extends SpecBase {
+
+  "CreateRecordSuccess Controller" - {
 
     "must return OK and the correct view for a GET" in {
 
-      val application = applicationBuilder(userAnswers = Some(emptyUserAnswers)).build()
+      val mockTraderProfileConnector: TraderProfileConnector = mock[TraderProfileConnector]
+      when(mockTraderProfileConnector.checkTraderProfile(any())(any())) thenReturn Future.successful(true)
+
+      val application = applicationBuilder(userAnswers = Some(emptyUserAnswers))
+        .overrides(inject.bind[TraderProfileConnector].toInstance(mockTraderProfileConnector))
+        .build()
 
       running(application) {
-        val request = FakeRequest(GET, routes.UkimsNumberChangeController.onPageLoad().url)
+        val request = FakeRequest(GET, routes.CreateProfileSuccessController.onPageLoad().url)
 
         val result = route(application, request).value
 
-        val view = application.injector.instanceOf[UkimsNumberChangeView]
+        val view = application.injector.instanceOf[CreateProfileSuccessView]
 
         status(result) mustEqual OK
         contentAsString(result) mustEqual view()(request, messages(application)).toString
