+++ test/controllers/profile/NirmsNumberControllerSpec.scala
@@ -14,34 +14,33 @@
  * limitations under the License.
  */
 
-package controllers
+package controllers.profile
 
 import base.SpecBase
 import base.TestConstants.{testEori, userAnswersId}
 import connectors.TraderProfileConnector
-import forms.NiphlNumberFormProvider
+import forms.NirmsNumberFormProvider
 import models.{NormalMode, TraderProfile, UserAnswers}
 import navigation.{FakeNavigator, Navigator}
-import org.apache.pekko.Done
 import org.mockito.ArgumentMatchers.{any, eq => eqTo}
 import org.mockito.Mockito.{never, verify, when}
 import org.scalatestplus.mockito.MockitoSugar
-import pages.{HasNiphlUpdatePage, NiphlNumberPage, NiphlNumberUpdatePage}
+import pages.{HasNirmsUpdatePage, NirmsNumberPage, NirmsNumberUpdatePage}
 import play.api.inject.bind
 import play.api.mvc.Call
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
 import repositories.SessionRepository
 import services.AuditService
-import views.html.NiphlNumberView
+import uk.gov.hmrc.play.bootstrap.binders.RedirectUrl
 
 import scala.concurrent.Future
 
-class NiphlNumberControllerSpec extends SpecBase with MockitoSugar {
+class NirmsNumberControllerSpec extends SpecBase with MockitoSugar {
 
   private def onwardRoute = Call("GET", "/foo")
 
-  val formProvider = new NiphlNumberFormProvider()
+  val formProvider = new NirmsNumberFormProvider()
   private val form = formProvider()
 
   val mockTraderProfileConnector: TraderProfileConnector = mock[TraderProfileConnector]
@@ -50,13 +49,11 @@ class NiphlNumberControllerSpec extends SpecBase with MockitoSugar {
 
   when(mockTraderProfileConnector.checkTraderProfile(any())(any())) thenReturn Future.successful(false)
 
-  private lazy val niphlNumberRouteCreate = routes.NiphlNumberController.onPageLoadCreate(NormalMode).url
-
-  private lazy val niphlNumberRouteUpdate = routes.NiphlNumberController.onPageLoadUpdate(NormalMode).url
-
-  "NiphlNumber Controller" - {
+  "NirmsNumber Controller" - {
 
     ".create" - {
+
+      val nirmsNumberRoute = routes.NirmsNumberController.onPageLoadCreate(NormalMode).url
 
       "must return OK and the correct view for a GET" in {
 
@@ -67,14 +64,14 @@ class NiphlNumberControllerSpec extends SpecBase with MockitoSugar {
           .build()
 
         running(application) {
-          val request = FakeRequest(GET, niphlNumberRouteCreate)
+          val request = FakeRequest(GET, nirmsNumberRoute)
 
           val result = route(application, request).value
 
-          val view = application.injector.instanceOf[NiphlNumberView]
+          val view = application.injector.instanceOf[NirmsNumberView]
 
           status(result) mustEqual OK
-          contentAsString(result) mustEqual view(form, routes.NiphlNumberController.onSubmitCreate(NormalMode))(
+          contentAsString(result) mustEqual view(form, routes.NirmsNumberController.onSubmitCreate(NormalMode))(
             request,
             messages(application)
           ).toString
@@ -83,7 +80,7 @@ class NiphlNumberControllerSpec extends SpecBase with MockitoSugar {
 
       "must populate the view correctly on a GET when the question has previously been answered" in {
 
-        val userAnswers = UserAnswers(userAnswersId).set(NiphlNumberPage, "answer").success.value
+        val userAnswers = UserAnswers(userAnswersId).set(NirmsNumberPage, "answer").success.value
 
         val application = applicationBuilder(userAnswers = Some(userAnswers))
           .overrides(
@@ -92,16 +89,16 @@ class NiphlNumberControllerSpec extends SpecBase with MockitoSugar {
           .build()
 
         running(application) {
-          val request = FakeRequest(GET, niphlNumberRouteCreate)
+          val request = FakeRequest(GET, nirmsNumberRoute)
 
-          val view = application.injector.instanceOf[NiphlNumberView]
+          val view = application.injector.instanceOf[NirmsNumberView]
 
           val result = route(application, request).value
 
           status(result) mustEqual OK
           contentAsString(result) mustEqual view(
             form.fill("answer"),
-            routes.NiphlNumberController.onSubmitCreate(NormalMode)
+            routes.NirmsNumberController.onSubmitCreate(NormalMode)
           )(request, messages(application)).toString
         }
       }
@@ -120,8 +117,8 @@ class NiphlNumberControllerSpec extends SpecBase with MockitoSugar {
 
         running(application) {
           val request =
-            FakeRequest(POST, niphlNumberRouteCreate)
-              .withFormUrlEncodedBody(("value", "SN12345"))
+            FakeRequest(POST, nirmsNumberRoute)
+              .withFormUrlEncodedBody(("value", "RMS-GB-123456"))
 
           val result = route(application, request).value
 
@@ -136,17 +133,17 @@ class NiphlNumberControllerSpec extends SpecBase with MockitoSugar {
 
         running(application) {
           val request =
-            FakeRequest(POST, niphlNumberRouteCreate)
+            FakeRequest(POST, nirmsNumberRoute)
               .withFormUrlEncodedBody(("value", ""))
 
           val boundForm = form.bind(Map("value" -> ""))
 
-          val view = application.injector.instanceOf[NiphlNumberView]
+          val view = application.injector.instanceOf[NirmsNumberView]
 
           val result = route(application, request).value
 
           status(result) mustEqual BAD_REQUEST
-          contentAsString(result) mustEqual view(boundForm, routes.NiphlNumberController.onSubmitCreate(NormalMode))(
+          contentAsString(result) mustEqual view(boundForm, routes.NirmsNumberController.onSubmitCreate(NormalMode))(
             request,
             messages(application)
           ).toString
@@ -162,7 +159,7 @@ class NiphlNumberControllerSpec extends SpecBase with MockitoSugar {
           .build()
 
         running(application) {
-          val request = FakeRequest(GET, niphlNumberRouteCreate)
+          val request = FakeRequest(GET, nirmsNumberRoute)
 
           val result = route(application, request).value
 
@@ -182,7 +179,7 @@ class NiphlNumberControllerSpec extends SpecBase with MockitoSugar {
           .build()
 
         running(application) {
-          val request = FakeRequest(GET, niphlNumberRouteCreate)
+          val request = FakeRequest(GET, nirmsNumberRoute)
 
           val result = route(application, request).value
 
@@ -197,7 +194,7 @@ class NiphlNumberControllerSpec extends SpecBase with MockitoSugar {
 
         running(application) {
           val request =
-            FakeRequest(POST, niphlNumberRouteCreate)
+            FakeRequest(POST, nirmsNumberRoute)
               .withFormUrlEncodedBody(("value", "answer"))
 
           val result = route(application, request).value
@@ -210,7 +207,9 @@ class NiphlNumberControllerSpec extends SpecBase with MockitoSugar {
 
     ".update" - {
 
-      "must return OK and the correct view for a GET when HasNiphl hasn't been answered when there is a niphl number" in {
+      val nirmsNumberRoute = routes.NirmsNumberController.onPageLoadUpdate(NormalMode).url
+
+      "must return OK and the correct view for a GET when HasNirms hasn't been answered when there is a nirms number" in {
 
         val traderProfile    = TraderProfile(testEori, "1", Some("2"), Some("3"), false)
         val mockAuditService = mock[AuditService]
@@ -232,16 +231,16 @@ class NiphlNumberControllerSpec extends SpecBase with MockitoSugar {
           .build()
 
         running(application) {
-          val request = FakeRequest(GET, niphlNumberRouteUpdate)
+          val request = FakeRequest(GET, nirmsNumberRoute)
 
           val result = route(application, request).value
 
-          val view = application.injector.instanceOf[NiphlNumberView]
+          val view = application.injector.instanceOf[NirmsNumberView]
 
           status(result) mustEqual OK
           contentAsString(result) mustEqual view(
-            form.fill("3"),
-            routes.NiphlNumberController.onSubmitUpdate(NormalMode)
+            form.fill("2"),
+            routes.NirmsNumberController.onSubmitUpdate(NormalMode)
           )(
             request,
             messages(application)
@@ -253,9 +252,158 @@ class NiphlNumberControllerSpec extends SpecBase with MockitoSugar {
         }
       }
 
-      "must return OK and the correct view for a GET when HasNiphl hasn't been answered when there isn't a niphl number" in {
+      "must return OK and the correct view for a GET when HasNirms has been answered when there is a nirms number" in {
 
-        val traderProfile    = TraderProfile(testEori, "1", Some("2"), None, eoriChanged = false)
+        val traderProfile = TraderProfile(testEori, "1", Some("2"), Some("3"), false)
+
+        val mockAuditService = mock[AuditService]
+
+        when(mockTraderProfileConnector.getTraderProfile(eqTo(testEori))(any())) thenReturn Future.successful(
+          traderProfile
+        )
+
+        when(mockSessionRepository.set(any())) thenReturn Future.successful(
+          true
+        )
+
+        val application =
+          applicationBuilder(userAnswers = Some(emptyUserAnswers.set(HasNirmsUpdatePage, true).success.value))
+            .overrides(
+              bind[TraderProfileConnector].toInstance(mockTraderProfileConnector),
+              bind[SessionRepository].toInstance(mockSessionRepository),
+              bind[AuditService].toInstance(mockAuditService)
+            )
+            .build()
+
+        running(application) {
+          val request = FakeRequest(GET, nirmsNumberRoute)
+
+          val result = route(application, request).value
+
+          val view = application.injector.instanceOf[NirmsNumberView]
+
+          status(result) mustEqual OK
+          contentAsString(result) mustEqual view(
+            form.fill("2"),
+            routes.NirmsNumberController.onSubmitUpdate(NormalMode)
+          )(
+            request,
+            messages(application)
+          ).toString
+
+          withClue("must not try and submit an audit") {
+            verify(mockAuditService, never()).auditMaintainProfile(any(), any(), any())(any())
+          }
+        }
+      }
+
+      "must return OK and the correct view for a GET when HasNirms has been answered when there isn't a nirms number" in {
+
+        val traderProfile = TraderProfile(testEori, "1", None, Some("3"), eoriChanged = false)
+
+        val mockAuditService = mock[AuditService]
+
+        when(mockTraderProfileConnector.getTraderProfile(eqTo(testEori))(any())) thenReturn Future.successful(
+          traderProfile
+        )
+
+        when(mockSessionRepository.set(any())) thenReturn Future.successful(
+          true
+        )
+
+        val application =
+          applicationBuilder(userAnswers = Some(emptyUserAnswers.set(HasNirmsUpdatePage, true).success.value))
+            .overrides(
+              bind[TraderProfileConnector].toInstance(mockTraderProfileConnector),
+              bind[SessionRepository].toInstance(mockSessionRepository),
+              bind[AuditService].toInstance(mockAuditService)
+            )
+            .build()
+
+        running(application) {
+          val request = FakeRequest(GET, nirmsNumberRoute)
+
+          val result = route(application, request).value
+
+          val view = application.injector.instanceOf[NirmsNumberView]
+
+          status(result) mustEqual OK
+          contentAsString(result) mustEqual view(form, routes.NirmsNumberController.onSubmitUpdate(NormalMode))(
+            request,
+            messages(application)
+          ).toString
+
+          withClue("must not try and submit an audit") {
+            verify(mockAuditService, never()).auditMaintainProfile(any(), any(), any())(any())
+          }
+        }
+      }
+
+      "must redirect to CyaMaintainProfile for a POST" in {
+        val answer = "RMS-GB-123456"
+
+        val traderProfile = TraderProfile(testEori, "1", Some(answer), Some("3"), eoriChanged = false)
+
+        val userAnswers = emptyUserAnswers
+          .set(HasNirmsUpdatePage, true)
+          .success
+          .value
+          .set(NirmsNumberUpdatePage, answer)
+          .success
+          .value
+
+        when(mockSessionRepository.set(any())) thenReturn Future.successful(true)
+
+        val mockTraderProfileConnector: TraderProfileConnector = mock[TraderProfileConnector]
+
+        when(mockTraderProfileConnector.getTraderProfile(any())(any())) thenReturn Future.successful(traderProfile)
+
+        val application =
+          applicationBuilder(userAnswers = Some(userAnswers))
+            .overrides(
+              bind[SessionRepository].toInstance(mockSessionRepository),
+              bind[TraderProfileConnector].toInstance(mockTraderProfileConnector)
+            )
+            .build()
+
+        running(application) {
+          val request =
+            FakeRequest(POST, nirmsNumberRoute)
+              .withFormUrlEncodedBody(("value", answer))
+
+          val result = route(application, request).value
+
+          status(result) mustEqual SEE_OTHER
+          redirectLocation(result).value mustEqual routes.CyaMaintainProfileController.onPageLoadNirmsNumber().url
+        }
+      }
+
+      "must return a Bad Request and errors when invalid data is submitted" in {
+
+        val application = applicationBuilder(userAnswers = Some(emptyUserAnswers)).build()
+
+        running(application) {
+          val request =
+            FakeRequest(POST, nirmsNumberRoute)
+              .withFormUrlEncodedBody(("value", ""))
+
+          val boundForm = form.bind(Map("value" -> ""))
+
+          val view = application.injector.instanceOf[NirmsNumberView]
+
+          val result = route(application, request).value
+
+          status(result) mustEqual BAD_REQUEST
+          contentAsString(result) mustEqual view(boundForm, routes.NirmsNumberController.onSubmitUpdate(NormalMode))(
+            request,
+            messages(application)
+          ).toString
+        }
+      }
+
+      "must redirect to Journey Recovery for a GET if HasNirms hasn't been answered and trader profile has no nirms number" in {
+
+        val traderProfile    = TraderProfile(testEori, "1", None, Some("3"), eoriChanged = false)
         val mockAuditService = mock[AuditService]
 
         when(mockTraderProfileConnector.getTraderProfile(eqTo(testEori))(any())) thenReturn Future.successful(
@@ -274,29 +422,26 @@ class NiphlNumberControllerSpec extends SpecBase with MockitoSugar {
           )
           .build()
 
+        val expectedRedirectLocation =
+          routes.JourneyRecoveryController
+            .onPageLoad(
+              Some(RedirectUrl(routes.ProfileController.onPageLoad().url))
+            )
+            .url
+
         running(application) {
-          val request = FakeRequest(GET, niphlNumberRouteUpdate)
+          val request = FakeRequest(GET, nirmsNumberRoute)
 
           val result = route(application, request).value
 
-          val view = application.injector.instanceOf[NiphlNumberView]
-
-          status(result) mustEqual OK
-          contentAsString(result) mustEqual view(form, routes.NiphlNumberController.onSubmitUpdate(NormalMode))(
-            request,
-            messages(application)
-          ).toString
-
-          withClue("must not try and submit an audit") {
-            verify(mockAuditService, never()).auditMaintainProfile(any(), any(), any())(any())
-          }
+          status(result) mustEqual SEE_OTHER
+          redirectLocation(result).value mustEqual expectedRedirectLocation
         }
       }
 
-      "must return OK and the correct view for a GET when HasNiphl has been answered when there is a niphl number" in {
+      "must redirect to Journey Recovery for a GET if HasNirms has been answered no" in {
 
-        val traderProfile = TraderProfile(testEori, "1", Some("2"), Some("3"), false)
-
+        val traderProfile    = TraderProfile(testEori, "1", Some("2"), Some("3"), false)
         val mockAuditService = mock[AuditService]
 
         when(mockTraderProfileConnector.getTraderProfile(eqTo(testEori))(any())) thenReturn Future.successful(
@@ -307,190 +452,30 @@ class NiphlNumberControllerSpec extends SpecBase with MockitoSugar {
           true
         )
 
-        val application =
-          applicationBuilder(userAnswers = Some(emptyUserAnswers.set(HasNiphlUpdatePage, true).success.value))
-            .overrides(
-              bind[TraderProfileConnector].toInstance(mockTraderProfileConnector),
-              bind[SessionRepository].toInstance(mockSessionRepository),
-              bind[AuditService].toInstance(mockAuditService)
+        val userAnswers = emptyUserAnswers.set(HasNirmsUpdatePage, false).success.value
+
+        val application = applicationBuilder(userAnswers = Some(userAnswers))
+          .overrides(
+            bind[TraderProfileConnector].toInstance(mockTraderProfileConnector),
+            bind[SessionRepository].toInstance(mockSessionRepository),
+            bind[AuditService].toInstance(mockAuditService)
+          )
+          .build()
+
+        val expectedRedirectLocation =
+          routes.JourneyRecoveryController
+            .onPageLoad(
+              Some(RedirectUrl(routes.ProfileController.onPageLoad().url))
             )
-            .build()
+            .url
 
         running(application) {
-          val request = FakeRequest(GET, niphlNumberRouteUpdate)
-
-          val result = route(application, request).value
-
-          val view = application.injector.instanceOf[NiphlNumberView]
-
-          status(result) mustEqual OK
-          contentAsString(result) mustEqual view(
-            form.fill("3"),
-            routes.NiphlNumberController.onSubmitUpdate(NormalMode)
-          )(
-            request,
-            messages(application)
-          ).toString
-
-          withClue("must not try and submit an audit") {
-            verify(mockAuditService, never()).auditMaintainProfile(any(), any(), any())(any())
-          }
-        }
-      }
-
-      "must return OK and the correct view for a GET when HasNiphl has been answered when there isn't a niphl number" in {
-
-        val traderProfile = TraderProfile(testEori, "1", Some("3"), None, eoriChanged = false)
-
-        val mockAuditService = mock[AuditService]
-
-        when(mockTraderProfileConnector.getTraderProfile(eqTo(testEori))(any())) thenReturn Future.successful(
-          traderProfile
-        )
-
-        when(mockSessionRepository.set(any())) thenReturn Future.successful(
-          true
-        )
-
-        val application =
-          applicationBuilder(userAnswers = Some(emptyUserAnswers.set(HasNiphlUpdatePage, true).success.value))
-            .overrides(
-              bind[TraderProfileConnector].toInstance(mockTraderProfileConnector),
-              bind[SessionRepository].toInstance(mockSessionRepository),
-              bind[AuditService].toInstance(mockAuditService)
-            )
-            .build()
-
-        running(application) {
-          val request = FakeRequest(GET, niphlNumberRouteUpdate)
-
-          val result = route(application, request).value
-
-          val view = application.injector.instanceOf[NiphlNumberView]
-
-          status(result) mustEqual OK
-          contentAsString(result) mustEqual view(form, routes.NiphlNumberController.onSubmitUpdate(NormalMode))(
-            request,
-            messages(application)
-          ).toString
-
-          withClue("must not try and submit an audit") {
-            verify(mockAuditService, never()).auditMaintainProfile(any(), any(), any())(any())
-          }
-        }
-      }
-
-      "must redirect to Profile for a POST and submit data if value is different from original" in {
-        val answer = "SN12345"
-
-        val traderProfile = TraderProfile(testEori, "1", Some("2"), Some("3"), false)
-        val userAnswers   = emptyUserAnswers
-          .set(HasNiphlUpdatePage, true)
-          .success
-          .value
-          .set(NiphlNumberUpdatePage, answer)
-          .success
-          .value
-
-        val mockAuditService = mock[AuditService]
-
-        when(mockAuditService.auditMaintainProfile(any(), any(), any())(any))
-          .thenReturn(Future.successful(Done))
-
-        when(mockSessionRepository.set(any())) thenReturn Future.successful(true)
-
-        when(mockTraderProfileConnector.getTraderProfile(any())(any())) thenReturn Future.successful(traderProfile)
-
-        val application =
-          applicationBuilder(userAnswers = Some(userAnswers))
-            .overrides(
-              bind[Navigator].toInstance(new FakeNavigator(onwardRoute)),
-              bind[SessionRepository].toInstance(mockSessionRepository),
-              bind[TraderProfileConnector].toInstance(mockTraderProfileConnector),
-              bind[AuditService].toInstance(mockAuditService)
-            )
-            .build()
-
-        running(application) {
-          val request =
-            FakeRequest(POST, niphlNumberRouteUpdate)
-              .withFormUrlEncodedBody(("value", answer))
+          val request = FakeRequest(GET, nirmsNumberRoute)
 
           val result = route(application, request).value
 
           status(result) mustEqual SEE_OTHER
-          redirectLocation(result).value mustEqual onwardRoute.url
-        }
-      }
-
-      "must redirect to Profile for a POST and not submit data if value is the same as original" in {
-        val answer = "SN12345"
-
-        val traderProfile = TraderProfile(testEori, "1", Some("2"), Some(answer), eoriChanged = false)
-
-        val userAnswers = emptyUserAnswers
-          .set(HasNiphlUpdatePage, true)
-          .success
-          .value
-          .set(NiphlNumberUpdatePage, answer)
-          .success
-          .value
-
-        val mockAuditService = mock[AuditService]
-
-        when(mockSessionRepository.set(any())) thenReturn Future.successful(true)
-
-        val mockTraderProfileConnector: TraderProfileConnector = mock[TraderProfileConnector]
-
-        when(mockTraderProfileConnector.getTraderProfile(any())(any())) thenReturn Future.successful(traderProfile)
-
-        val application =
-          applicationBuilder(userAnswers = Some(userAnswers))
-            .overrides(
-              bind[SessionRepository].toInstance(mockSessionRepository),
-              bind[TraderProfileConnector].toInstance(mockTraderProfileConnector),
-              bind[AuditService].toInstance(mockAuditService)
-            )
-            .build()
-
-        running(application) {
-          val request =
-            FakeRequest(POST, niphlNumberRouteUpdate)
-              .withFormUrlEncodedBody(("value", answer))
-
-          val result = route(application, request).value
-
-          status(result) mustEqual SEE_OTHER
-          redirectLocation(result).value mustEqual routes.CyaMaintainProfileController.onPageLoadNiphlNumber().url
-          verify(mockTraderProfileConnector, never())
-            .submitTraderProfile(any(), any())(any())
-
-          withClue("must not try and submit an audit") {
-            verify(mockAuditService, never()).auditMaintainProfile(any(), any(), any())(any())
-          }
-        }
-      }
-
-      "must return a Bad Request and errors when invalid data is submitted" in {
-
-        val application = applicationBuilder(userAnswers = Some(emptyUserAnswers)).build()
-
-        running(application) {
-          val request =
-            FakeRequest(POST, niphlNumberRouteUpdate)
-              .withFormUrlEncodedBody(("value", ""))
-
-          val boundForm = form.bind(Map("value" -> ""))
-
-          val view = application.injector.instanceOf[NiphlNumberView]
-
-          val result = route(application, request).value
-
-          status(result) mustEqual BAD_REQUEST
-          contentAsString(result) mustEqual view(boundForm, routes.NiphlNumberController.onSubmitUpdate(NormalMode))(
-            request,
-            messages(application)
-          ).toString
+          redirectLocation(result).value mustEqual expectedRedirectLocation
         }
       }
 
@@ -500,7 +485,7 @@ class NiphlNumberControllerSpec extends SpecBase with MockitoSugar {
           .build()
 
         running(application) {
-          val request = FakeRequest(GET, niphlNumberRouteUpdate)
+          val request = FakeRequest(GET, nirmsNumberRoute)
 
           val result = route(application, request).value
 
@@ -515,7 +500,7 @@ class NiphlNumberControllerSpec extends SpecBase with MockitoSugar {
 
         running(application) {
           val request =
-            FakeRequest(POST, niphlNumberRouteUpdate)
+            FakeRequest(POST, nirmsNumberRoute)
               .withFormUrlEncodedBody(("value", "answer"))
 
           val result = route(application, request).value
@@ -524,6 +509,7 @@ class NiphlNumberControllerSpec extends SpecBase with MockitoSugar {
           redirectLocation(result).value mustEqual routes.JourneyRecoveryController.onPageLoad().url
         }
       }
+
     }
   }
 }
