+++ test/controllers/profile/UkimsNumberControllerSpec.scala
@@ -14,32 +14,32 @@
  * limitations under the License.
  */
 
-package controllers
+package controllers.profile
 
 import base.SpecBase
 import base.TestConstants.{testEori, userAnswersId}
 import connectors.TraderProfileConnector
-import forms.HasNiphlFormProvider
-import models.{NormalMode, TraderProfile, UserAnswers}
+import forms.UkimsNumberFormProvider
+import models.{CheckMode, NormalMode, TraderProfile, UserAnswers}
 import navigation.{FakeNavigator, Navigator}
 import org.mockito.ArgumentMatchers.{any, eq => eqTo}
-import org.mockito.Mockito.when
+import org.mockito.Mockito.{never, verify, when}
 import org.scalatestplus.mockito.MockitoSugar
-import pages.{HasNiphlPage, HasNiphlUpdatePage}
+import pages.{UkimsNumberPage, UkimsNumberUpdatePage}
 import play.api.inject.bind
 import play.api.mvc.Call
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
 import repositories.SessionRepository
-import views.html.HasNiphlView
+import services.AuditService
 
 import scala.concurrent.Future
 
-class HasNiphlControllerSpec extends SpecBase with MockitoSugar {
+class UkimsNumberControllerSpec extends SpecBase with MockitoSugar {
 
   private def onwardRoute = Call("GET", "/foo")
 
-  val formProvider = new HasNiphlFormProvider()
+  val formProvider = new UkimsNumberFormProvider()
   private val form = formProvider()
 
   val mockTraderProfileConnector: TraderProfileConnector = mock[TraderProfileConnector]
@@ -48,31 +48,34 @@ class HasNiphlControllerSpec extends SpecBase with MockitoSugar {
 
   when(mockTraderProfileConnector.checkTraderProfile(any())(any())) thenReturn Future.successful(false)
 
-  private lazy val hasNiphlRouteCreate = routes.HasNiphlController.onPageLoadCreate(NormalMode).url
-
-  private lazy val hasNiphlRouteUpdate = routes.HasNiphlController.onPageLoadUpdate(NormalMode).url
-
-  "HasNiphlController" - {
+  "UkimsNumberController" - {
 
     ".create" - {
+
+      val ukimsNumberRoute = routes.UkimsNumberController.onPageLoadCreate(NormalMode).url
 
       "must return OK and the correct view for a GET" in {
 
         val application = applicationBuilder(userAnswers = Some(emptyUserAnswers))
           .overrides(
+            bind[Navigator].toInstance(new FakeNavigator(onwardRoute)),
             bind[TraderProfileConnector].toInstance(mockTraderProfileConnector)
           )
           .build()
 
         running(application) {
-          val request = FakeRequest(GET, hasNiphlRouteCreate)
+          val request = FakeRequest(GET, ukimsNumberRoute)
 
           val result = route(application, request).value
 
-          val view = application.injector.instanceOf[HasNiphlView]
+          val view = application.injector.instanceOf[UkimsNumberView]
 
           status(result) mustEqual OK
-          contentAsString(result) mustEqual view(form, routes.HasNiphlController.onSubmitCreate(NormalMode))(
+          contentAsString(result) mustEqual view(
+            form,
+            routes.UkimsNumberController.onSubmitCreate(NormalMode),
+            isCreateJourney = true
+          )(
             request,
             messages(application)
           ).toString
@@ -81,23 +84,28 @@ class HasNiphlControllerSpec extends SpecBase with MockitoSugar {
 
       "must populate the view correctly on a GET when the question has previously been answered" in {
 
-        val userAnswers = UserAnswers(userAnswersId).set(HasNiphlPage, true).success.value
+        val userAnswers = UserAnswers(userAnswersId).set(UkimsNumberPage, "answer").success.value
 
         val application = applicationBuilder(userAnswers = Some(userAnswers))
           .overrides(
+            bind[Navigator].toInstance(new FakeNavigator(onwardRoute)),
             bind[TraderProfileConnector].toInstance(mockTraderProfileConnector)
           )
           .build()
 
         running(application) {
-          val request = FakeRequest(GET, hasNiphlRouteCreate)
+          val request = FakeRequest(GET, ukimsNumberRoute)
 
-          val view = application.injector.instanceOf[HasNiphlView]
+          val view = application.injector.instanceOf[UkimsNumberView]
 
           val result = route(application, request).value
 
           status(result) mustEqual OK
-          contentAsString(result) mustEqual view(form.fill(true), routes.HasNiphlController.onSubmitCreate(NormalMode))(
+          contentAsString(result) mustEqual view(
+            form.fill("answer"),
+            routes.UkimsNumberController.onSubmitCreate(NormalMode),
+            isCreateJourney = true
+          )(
             request,
             messages(application)
           ).toString
@@ -112,15 +120,14 @@ class HasNiphlControllerSpec extends SpecBase with MockitoSugar {
           applicationBuilder(userAnswers = Some(emptyUserAnswers))
             .overrides(
               bind[Navigator].toInstance(new FakeNavigator(onwardRoute)),
-              bind[SessionRepository].toInstance(mockSessionRepository),
-              bind[TraderProfileConnector].toInstance(mockTraderProfileConnector)
+              bind[SessionRepository].toInstance(mockSessionRepository)
             )
             .build()
 
         running(application) {
           val request =
-            FakeRequest(POST, hasNiphlRouteCreate)
-              .withFormUrlEncodedBody(("value", "true"))
+            FakeRequest(POST, ukimsNumberRoute)
+              .withFormUrlEncodedBody(("value", "XIUKIM47699357400020231115081800"))
 
           val result = route(application, request).value
 
@@ -131,25 +138,21 @@ class HasNiphlControllerSpec extends SpecBase with MockitoSugar {
 
       "must return a Bad Request and errors when invalid data is submitted" in {
 
-        val application = applicationBuilder(userAnswers = Some(emptyUserAnswers))
-          .overrides(
-            bind[TraderProfileConnector].toInstance(mockTraderProfileConnector)
-          )
-          .build()
+        val application = applicationBuilder(userAnswers = Some(emptyUserAnswers)).build()
 
         running(application) {
           val request =
-            FakeRequest(POST, hasNiphlRouteCreate)
+            FakeRequest(POST, ukimsNumberRoute)
               .withFormUrlEncodedBody(("value", ""))
 
           val boundForm = form.bind(Map("value" -> ""))
 
-          val view = application.injector.instanceOf[HasNiphlView]
+          val view = application.injector.instanceOf[UkimsNumberView]
 
           val result = route(application, request).value
 
           status(result) mustEqual BAD_REQUEST
-          contentAsString(result) mustEqual view(boundForm, routes.HasNiphlController.onSubmitCreate(NormalMode))(
+          contentAsString(result) mustEqual view(boundForm, routes.UkimsNumberController.onSubmitCreate(NormalMode))(
             request,
             messages(application)
           ).toString
@@ -165,7 +168,7 @@ class HasNiphlControllerSpec extends SpecBase with MockitoSugar {
           .build()
 
         running(application) {
-          val request = FakeRequest(GET, hasNiphlRouteCreate)
+          val request = FakeRequest(GET, ukimsNumberRoute)
 
           val result = route(application, request).value
 
@@ -176,8 +179,9 @@ class HasNiphlControllerSpec extends SpecBase with MockitoSugar {
 
       "must redirect to Homepage for a GET if profile already exists" in {
 
-        when(mockTraderProfileConnector.checkTraderProfile(any())(any())) thenReturn Future.successful(true)
+        val mockTraderProfileConnector: TraderProfileConnector = mock[TraderProfileConnector]
 
+        when(mockTraderProfileConnector.checkTraderProfile(any())(any())) thenReturn Future.successful(true)
         val application = applicationBuilder(userAnswers = None)
           .overrides(
             bind[TraderProfileConnector].toInstance(mockTraderProfileConnector)
@@ -185,7 +189,7 @@ class HasNiphlControllerSpec extends SpecBase with MockitoSugar {
           .build()
 
         running(application) {
-          val request = FakeRequest(GET, hasNiphlRouteCreate)
+          val request = FakeRequest(GET, ukimsNumberRoute)
 
           val result = route(application, request).value
 
@@ -196,16 +200,12 @@ class HasNiphlControllerSpec extends SpecBase with MockitoSugar {
 
       "must redirect to Journey Recovery for a POST if no existing data is found" in {
 
-        val application = applicationBuilder(userAnswers = None)
-          .overrides(
-            bind[TraderProfileConnector].toInstance(mockTraderProfileConnector)
-          )
-          .build()
+        val application = applicationBuilder(userAnswers = None).build()
 
         running(application) {
           val request =
-            FakeRequest(POST, hasNiphlRouteCreate)
-              .withFormUrlEncodedBody(("value", "true"))
+            FakeRequest(POST, ukimsNumberRoute)
+              .withFormUrlEncodedBody(("value", "answer"))
 
           val result = route(application, request).value
 
@@ -217,9 +217,12 @@ class HasNiphlControllerSpec extends SpecBase with MockitoSugar {
 
     ".update" - {
 
-      "must return OK and the correct view for a GET with saved answers" in {
+      val ukimsNumberRoute = routes.UkimsNumberController.onPageLoadUpdate(NormalMode).url
 
-        val traderProfile = TraderProfile(testEori, "1", Some("2"), Some("3"), false)
+      "must return OK and the correct view for a GET with all trader profile complete" in {
+
+        val traderProfile    = TraderProfile(testEori, "1", Some("2"), Some("3"), false)
+        val mockAuditService = mock[AuditService]
 
         when(mockTraderProfileConnector.getTraderProfile(eqTo(testEori))(any())) thenReturn Future.successful(
           traderProfile
@@ -232,100 +235,97 @@ class HasNiphlControllerSpec extends SpecBase with MockitoSugar {
         val application = applicationBuilder(userAnswers = Some(emptyUserAnswers))
           .overrides(
             bind[TraderProfileConnector].toInstance(mockTraderProfileConnector),
-            bind[SessionRepository].toInstance(mockSessionRepository)
+            bind[SessionRepository].toInstance(mockSessionRepository),
+            bind[AuditService].toInstance(mockAuditService)
           )
           .build()
 
         running(application) {
-          val request = FakeRequest(GET, hasNiphlRouteUpdate)
+          val request = FakeRequest(GET, ukimsNumberRoute)
 
           val result = route(application, request).value
 
-          val view = application.injector.instanceOf[HasNiphlView]
+          val view = application.injector.instanceOf[UkimsNumberView]
 
           status(result) mustEqual OK
-          contentAsString(result) mustEqual view(form.fill(true), routes.HasNiphlController.onSubmitUpdate(NormalMode))(
+          contentAsString(result) mustEqual view(
+            form.fill("1"),
+            routes.UkimsNumberController.onSubmitUpdate(NormalMode)
+          )(
             request,
             messages(application)
           ).toString
+
+          withClue("must not try and submit an audit") {
+            verify(mockAuditService, never()).auditMaintainProfile(any(), any(), any())(any())
+          }
         }
       }
 
-      "must populate the view correctly on a GET when the question has previously been answered" in {
+      "must return OK and the correct view for a GET with all trader profile complete if the user is returning from the CYA page and UKIMS number should be filled in" in {
 
-        val userAnswers = UserAnswers(userAnswersId).set(HasNiphlUpdatePage, true).success.value
+        val ukimsNumberCheckRoute = routes.UkimsNumberController.onPageLoadUpdate(CheckMode).url
+        val newUkims              = "newUkims"
+        val mockAuditService      = mock[AuditService]
+
+        val userAnswers = emptyUserAnswers.set(UkimsNumberUpdatePage, newUkims).success.value
+
+        when(mockSessionRepository.set(any())) thenReturn Future.successful(
+          true
+        )
 
         val application = applicationBuilder(userAnswers = Some(userAnswers))
           .overrides(
-            bind[TraderProfileConnector].toInstance(mockTraderProfileConnector)
+            bind[SessionRepository].toInstance(mockSessionRepository),
+            bind[AuditService].toInstance(mockAuditService)
           )
           .build()
 
         running(application) {
-          val request = FakeRequest(GET, hasNiphlRouteUpdate)
-
-          val view = application.injector.instanceOf[HasNiphlView]
+          val request = FakeRequest(GET, ukimsNumberCheckRoute)
 
           val result = route(application, request).value
 
+          val view = application.injector.instanceOf[UkimsNumberView]
+
           status(result) mustEqual OK
-          contentAsString(result) mustEqual view(form.fill(true), routes.HasNiphlController.onSubmitUpdate(NormalMode))(
+          contentAsString(result) mustEqual view(
+            form.fill(newUkims),
+            routes.UkimsNumberController.onSubmitUpdate(CheckMode)
+          )(
             request,
             messages(application)
           ).toString
+
+          withClue("must not try and submit an audit or get profile") {
+            verify(mockAuditService, never()).auditMaintainProfile(any(), any(), any())(any())
+          }
         }
       }
 
-      "must redirect to the next page when valid data is submitted" in {
+      "must redirect for a POST" in {
+
+        val answer = "XIUKIM47699357400020231115081800"
+
+        val userAnswers = emptyUserAnswers
+          .set(UkimsNumberUpdatePage, answer)
+          .success
+          .value
 
         when(mockSessionRepository.set(any())) thenReturn Future.successful(true)
 
-        val traderProfile = TraderProfile(testEori, "1", Some("2"), None, eoriChanged = false)
-
-        when(mockTraderProfileConnector.getTraderProfile(any())(any())) thenReturn Future.successful(traderProfile)
-
         val application =
-          applicationBuilder(userAnswers = Some(emptyUserAnswers))
+          applicationBuilder(userAnswers = Some(userAnswers))
             .overrides(
               bind[Navigator].toInstance(new FakeNavigator(onwardRoute)),
-              bind[SessionRepository].toInstance(mockSessionRepository),
-              bind[TraderProfileConnector].toInstance(mockTraderProfileConnector)
+              bind[SessionRepository].toInstance(mockSessionRepository)
             )
             .build()
 
         running(application) {
           val request =
-            FakeRequest(POST, hasNiphlRouteUpdate)
-              .withFormUrlEncodedBody(("value", "true"))
-
-          val result = route(application, request).value
-
-          status(result) mustEqual SEE_OTHER
-          redirectLocation(result).value mustEqual onwardRoute.url
-        }
-      }
-
-      "must redirect to ProfilePage when no changes made" in {
-
-        when(mockSessionRepository.set(any())) thenReturn Future.successful(true)
-
-        val traderProfile = TraderProfile(testEori, "1", Some("2"), Some("3"), false)
-
-        when(mockTraderProfileConnector.getTraderProfile(any())(any())) thenReturn Future.successful(traderProfile)
-
-        val application =
-          applicationBuilder(userAnswers = Some(emptyUserAnswers))
-            .overrides(
-              bind[SessionRepository].toInstance(mockSessionRepository),
-              bind[TraderProfileConnector].toInstance(mockTraderProfileConnector),
-              bind[Navigator].toInstance(new FakeNavigator(onwardRoute))
-            )
-            .build()
-
-        running(application) {
-          val request =
-            FakeRequest(POST, hasNiphlRouteUpdate)
-              .withFormUrlEncodedBody(("value", "true"))
+            FakeRequest(POST, ukimsNumberRoute)
+              .withFormUrlEncodedBody(("value", answer))
 
           val result = route(application, request).value
 
@@ -336,22 +336,21 @@ class HasNiphlControllerSpec extends SpecBase with MockitoSugar {
 
       "must return a Bad Request and errors when invalid data is submitted" in {
 
-        val application = applicationBuilder(userAnswers = Some(emptyUserAnswers))
-          .build()
+        val application = applicationBuilder(userAnswers = Some(emptyUserAnswers)).build()
 
         running(application) {
           val request =
-            FakeRequest(POST, hasNiphlRouteUpdate)
+            FakeRequest(POST, ukimsNumberRoute)
               .withFormUrlEncodedBody(("value", ""))
 
           val boundForm = form.bind(Map("value" -> ""))
 
-          val view = application.injector.instanceOf[HasNiphlView]
+          val view = application.injector.instanceOf[UkimsNumberView]
 
           val result = route(application, request).value
 
           status(result) mustEqual BAD_REQUEST
-          contentAsString(result) mustEqual view(boundForm, routes.HasNiphlController.onSubmitUpdate(NormalMode))(
+          contentAsString(result) mustEqual view(boundForm, routes.UkimsNumberController.onSubmitUpdate(NormalMode))(
             request,
             messages(application)
           ).toString
@@ -364,7 +363,7 @@ class HasNiphlControllerSpec extends SpecBase with MockitoSugar {
           .build()
 
         running(application) {
-          val request = FakeRequest(GET, hasNiphlRouteUpdate)
+          val request = FakeRequest(GET, ukimsNumberRoute)
 
           val result = route(application, request).value
 
@@ -375,13 +374,12 @@ class HasNiphlControllerSpec extends SpecBase with MockitoSugar {
 
       "must redirect to Journey Recovery for a POST if no existing data is found" in {
 
-        val application = applicationBuilder(userAnswers = None)
-          .build()
+        val application = applicationBuilder(userAnswers = None).build()
 
         running(application) {
           val request =
-            FakeRequest(POST, hasNiphlRouteUpdate)
-              .withFormUrlEncodedBody(("value", "true"))
+            FakeRequest(POST, ukimsNumberRoute)
+              .withFormUrlEncodedBody(("value", "answer"))
 
           val result = route(application, request).value
