+++ test/navigation/NavigatorSpec.scala
@@ -138,405 +138,6 @@ class NavigatorSpec extends SpecBase with BeforeAndAfterEach {
         }
       }
 
-      "in Create Profile Journey" - {
-
-        "must go from ProfileSetupPage" - {
-
-          "to UseExistingUkimsNumber when historic data" in {
-
-            val userAnswers = emptyUserAnswers
-              .set(
-                HistoricProfileDataQuery,
-                HistoricProfileData("GB123456789", "GB123456789", Some("XIUKIMS1234567890"), None, None)
-              )
-              .success
-              .value
-
-            navigator.nextPage(ProfileSetupPage, NormalMode, userAnswers) mustBe routes.UseExistingUkimsNumberController
-              .onPageLoad()
-          }
-
-          "to UkimsNumberPage when no historic data" in {
-
-            navigator.nextPage(ProfileSetupPage, NormalMode, emptyUserAnswers) mustBe routes.UkimsNumberController
-              .onPageLoadCreate(NormalMode)
-          }
-        }
-
-        "must go from UkimsNumberPage to HasNirmsPage" in {
-
-          navigator.nextPage(UkimsNumberPage, NormalMode, emptyUserAnswers) mustBe routes.HasNirmsController
-            .onPageLoadCreate(
-              NormalMode
-            )
-        }
-
-        "must go from UseExistingNirmsPage" - {
-
-          "to HasNirmsPage when answer is Yes" in {
-
-            val answers = UserAnswers(userAnswersId).set(UseExistingUkimsNumberPage, true).success.value
-            navigator.nextPage(UseExistingUkimsNumberPage, NormalMode, answers) mustBe routes.HasNirmsController
-              .onPageLoadCreate(
-                NormalMode
-              )
-          }
-
-          "to UkimsNumberController when answer is No" in {
-
-            val answers = UserAnswers(userAnswersId).set(UseExistingUkimsNumberPage, false).success.value
-            navigator.nextPage(UseExistingUkimsNumberPage, NormalMode, answers) mustBe routes.UkimsNumberController
-              .onPageLoadCreate(
-                NormalMode
-              )
-          }
-
-          "to JourneyRecoveryPage when answer is not present" in {
-
-            navigator.nextPage(
-              UseExistingUkimsNumberPage,
-              NormalMode,
-              emptyUserAnswers
-            ) mustBe routes.JourneyRecoveryController
-              .onPageLoad()
-          }
-        }
-
-        "must go from HasNirmsPage" - {
-
-          "to NirmsNumberPage when answer is Yes" in {
-
-            val answers = UserAnswers(userAnswersId).set(HasNirmsPage, true).success.value
-            navigator.nextPage(HasNirmsPage, NormalMode, answers) mustBe routes.NirmsNumberController.onPageLoadCreate(
-              NormalMode
-            )
-          }
-
-          "to HasNiphlPage when answer is No" in {
-
-            val answers = UserAnswers(userAnswersId).set(HasNirmsPage, false).success.value
-            navigator.nextPage(HasNirmsPage, NormalMode, answers) mustBe routes.HasNiphlController.onPageLoadCreate(
-              NormalMode
-            )
-          }
-
-          "to JourneyRecoveryPage when answer is not present" in {
-
-            navigator.nextPage(
-              HasNirmsPage,
-              NormalMode,
-              emptyUserAnswers
-            ) mustBe routes.JourneyRecoveryController
-              .onPageLoad()
-          }
-        }
-
-        "must go from NirmsNumberPage to HasNiphlPage" in {
-
-          navigator.nextPage(NirmsNumberPage, NormalMode, emptyUserAnswers) mustBe routes.HasNiphlController
-            .onPageLoadCreate(
-              NormalMode
-            )
-        }
-
-        "must go from HasNiphlPage" - {
-
-          "to NiphlNumberPage when answer is Yes" in {
-
-            val answers = UserAnswers(userAnswersId).set(HasNiphlPage, true).success.value
-            navigator.nextPage(HasNiphlPage, NormalMode, answers) mustBe routes.NiphlNumberController.onPageLoadCreate(
-              NormalMode
-            )
-          }
-
-          "to CyaCreateProfile when answer is No" in {
-
-            val answers = UserAnswers(userAnswersId).set(HasNiphlPage, false).success.value
-            navigator.nextPage(HasNiphlPage, NormalMode, answers) mustBe routes.CyaCreateProfileController.onPageLoad()
-          }
-
-          "to JourneyRecoveryPage when answer is not present" in {
-
-            navigator.nextPage(
-              HasNiphlPage,
-              NormalMode,
-              emptyUserAnswers
-            ) mustBe routes.JourneyRecoveryController
-              .onPageLoad()
-          }
-        }
-
-        "must go from NiphlNumberPage to CyaCreateProfile" in {
-
-          navigator.nextPage(
-            NiphlNumberPage,
-            NormalMode,
-            emptyUserAnswers
-          ) mustBe routes.CyaCreateProfileController.onPageLoad()
-        }
-
-        "must go from CyaCreateProfile to CreateProfileSuccess" in {
-
-          navigator.nextPage(
-            CyaCreateProfilePage,
-            NormalMode,
-            emptyUserAnswers
-          ) mustBe routes.CreateProfileSuccessController.onPageLoad()
-        }
-      }
-
-      "in Update Profile Journey" - {
-
-        "must go from UkimsNumberUpdatePage to CyaMaintainProfilePage" in {
-
-          navigator.nextPage(
-            UkimsNumberUpdatePage,
-            NormalMode,
-            emptyUserAnswers
-          ) mustBe routes.CyaMaintainProfileController.onPageLoadUkimsNumber()
-        }
-
-        "must go from HasNirmsUpdatePage" - {
-
-          "to NirmsNumberUpdatePage when answer is Yes" in {
-
-            val answers = UserAnswers(userAnswersId).set(HasNirmsUpdatePage, true).success.value
-            navigator.nextPage(
-              HasNirmsUpdatePage,
-              NormalMode,
-              answers
-            ) mustBe routes.NirmsNumberController.onPageLoadUpdate(NormalMode)
-          }
-
-          "to RemoveNirmsPage when answer is No and Nirms number is associated to profile" in {
-
-            val answers = UserAnswers(userAnswersId)
-              .set(HasNirmsUpdatePage, false)
-              .success
-              .value
-              .set(
-                TraderProfileQuery,
-                TraderProfile("actorId", "ukims", Some("nirms"), Some("niphl"), eoriChanged = false)
-              )
-              .success
-              .value
-
-            navigator.nextPage(HasNirmsUpdatePage, NormalMode, answers) mustBe routes.RemoveNirmsController
-              .onPageLoad()
-          }
-
-          "to RemoveNirmsPage when answer is No and Nirms number is not associated to profile" in {
-
-            val answers = UserAnswers(userAnswersId)
-              .set(HasNirmsUpdatePage, false)
-              .success
-              .value
-              .set(TraderProfileQuery, TraderProfile("actorId", "ukims", None, Some("niphl"), eoriChanged = false))
-              .success
-              .value
-
-            navigator.nextPage(
-              HasNirmsUpdatePage,
-              NormalMode,
-              answers
-            ) mustBe routes.CyaMaintainProfileController.onPageLoadNirms()
-          }
-
-          "to JourneyRecoveryPage when answer is not present" in {
-            val continueUrl = RedirectUrl(routes.ProfileController.onPageLoad().url)
-
-            navigator.nextPage(
-              HasNirmsUpdatePage,
-              NormalMode,
-              emptyUserAnswers
-            ) mustBe routes.JourneyRecoveryController
-              .onPageLoad(Some(continueUrl))
-          }
-
-          "to JourneyRecoveryPage when TraderProfileQuery not present" in {
-            val answers = UserAnswers(userAnswersId)
-              .set(HasNirmsUpdatePage, false)
-              .success
-              .value
-
-            val continueUrl = RedirectUrl(routes.ProfileController.onPageLoad().url)
-
-            navigator.nextPage(
-              HasNirmsUpdatePage,
-              NormalMode,
-              answers
-            ) mustBe routes.JourneyRecoveryController
-              .onPageLoad(Some(continueUrl))
-          }
-        }
-
-        "must go from RemoveNirmsPage" - {
-
-          "to CyaMaintainProfile when user answered No" in {
-
-            val answers = UserAnswers(userAnswersId).set(RemoveNirmsPage, false).success.value
-
-            navigator.nextPage(
-              RemoveNirmsPage,
-              NormalMode,
-              answers
-            ) mustBe routes.CyaMaintainProfileController.onPageLoadNirmsNumber()
-          }
-
-          "to Cya NIRMS registered when user answered yes" in {
-
-            val answers = UserAnswers(userAnswersId).set(RemoveNirmsPage, true).success.value
-
-            navigator.nextPage(
-              RemoveNirmsPage,
-              NormalMode,
-              answers
-            ) mustBe routes.CyaMaintainProfileController.onPageLoadNirms()
-          }
-
-          "to ProfilePage when answer is not present" in {
-
-            navigator.nextPage(
-              RemoveNirmsPage,
-              NormalMode,
-              emptyUserAnswers
-            ) mustBe routes.ProfileController.onPageLoad()
-          }
-        }
-
-        "must go from NirmsNumberUpdatePage to CyaMaintainProfile" in {
-
-          navigator.nextPage(
-            NirmsNumberUpdatePage,
-            NormalMode,
-            emptyUserAnswers
-          ) mustBe routes.CyaMaintainProfileController.onPageLoadNirmsNumber()
-        }
-
-        "must go from CyaMaintainProfilePage to CyaMaintainProfile" in {
-
-          navigator.nextPage(
-            CyaMaintainProfilePage,
-            NormalMode,
-            emptyUserAnswers
-          ) mustBe routes.ProfileController.onPageLoad()
-        }
-
-        "must go from HasNiphlUpdatePage" - {
-
-          "to NiphlNumberUpdatePage when answer is Yes" in {
-
-            val answers = UserAnswers(userAnswersId).set(HasNiphlUpdatePage, true).success.value
-            navigator.nextPage(
-              HasNiphlUpdatePage,
-              NormalMode,
-              answers
-            ) mustBe routes.NiphlNumberController.onPageLoadUpdate(NormalMode)
-          }
-
-          "to RemoveNiphlPage when answer is No and Niphl number is associated to profile" in {
-
-            val answers = UserAnswers(userAnswersId)
-              .set(HasNiphlUpdatePage, false)
-              .success
-              .value
-              .set(
-                TraderProfileQuery,
-                TraderProfile("actorId", "ukims", Some("nirms"), Some("niphl"), eoriChanged = false)
-              )
-              .success
-              .value
-
-            navigator.nextPage(HasNiphlUpdatePage, NormalMode, answers) mustBe routes.RemoveNiphlController.onPageLoad()
-          }
-
-          "to RemoveNiphlPage when answer is No and Niphl number is not associated to profile" in {
-
-            val answers = UserAnswers(userAnswersId)
-              .set(HasNiphlUpdatePage, false)
-              .success
-              .value
-              .set(TraderProfileQuery, TraderProfile("actorId", "ukims", Some("nirms"), None, eoriChanged = false))
-              .success
-              .value
-
-            navigator.nextPage(
-              HasNiphlUpdatePage,
-              NormalMode,
-              answers
-            ) mustBe routes.CyaMaintainProfileController.onPageLoadNiphl()
-          }
-
-          "to JourneyRecoveryPage when answer is not present" in {
-            val continueUrl = RedirectUrl(routes.ProfileController.onPageLoad().url)
-
-            navigator.nextPage(
-              HasNiphlUpdatePage,
-              NormalMode,
-              emptyUserAnswers
-            ) mustBe routes.JourneyRecoveryController
-              .onPageLoad(Some(continueUrl))
-          }
-
-          "to JourneyRecoveryPage when TraderProfileQuery not present" in {
-            val answers = UserAnswers(userAnswersId)
-              .set(HasNiphlUpdatePage, false)
-              .success
-              .value
-
-            val continueUrl = RedirectUrl(routes.ProfileController.onPageLoad().url)
-
-            navigator.nextPage(
-              HasNiphlUpdatePage,
-              NormalMode,
-              answers
-            ) mustBe routes.JourneyRecoveryController
-              .onPageLoad(Some(continueUrl))
-          }
-        }
-
-        "must go from RemoveNiphlPage" - {
-          "to Check Your Answers for Niphls Number when user answered No" in {
-
-            val answers = UserAnswers(userAnswersId).set(RemoveNiphlPage, false).success.value
-            navigator.nextPage(
-              RemoveNiphlPage,
-              NormalMode,
-              answers
-            ) mustBe routes.CyaMaintainProfileController.onPageLoadNiphlNumber()
-          }
-
-          "to Check Your Answers for NIPHL registered when user answered yes" in {
-
-            val answers = UserAnswers(userAnswersId).set(RemoveNiphlPage, true).success.value
-            navigator.nextPage(
-              RemoveNiphlPage,
-              NormalMode,
-              answers
-            ) mustBe routes.CyaMaintainProfileController.onPageLoadNiphl()
-          }
-
-          "to ProfilePage when answer is not present" in {
-
-            navigator.nextPage(
-              RemoveNiphlPage,
-              NormalMode,
-              emptyUserAnswers
-            ) mustBe routes.ProfileController.onPageLoad()
-          }
-        }
-
-        "must go from NiphlNumberUpdatePage to ProfilePage" in {
-
-          navigator.nextPage(
-            NiphlNumberUpdatePage,
-            NormalMode,
-            emptyUserAnswers
-          ) mustBe routes.CyaMaintainProfileController.onPageLoadNiphlNumber()
-        }
-
-      }
-
       "in Create Record Journey" - {
 
         "must go from CreateRecordStartPage to TraderReferencePage" in {
@@ -3022,6 +2623,7 @@ class NavigatorSpec extends SpecBase with BeforeAndAfterEach {
         }
 
       }
+
       "in Supplementary Unit Update Journey" - {
 
         "must go from HasSupplementaryUnitUpdatePage" - {
@@ -3098,19 +2700,6 @@ class NavigatorSpec extends SpecBase with BeforeAndAfterEach {
             NormalMode,
             emptyUserAnswers
           ) mustBe routes.DownloadRequestSuccessController.onPageLoad()
-        }
-      }
-
-      "in the new UKIMS number update journey" - {
-
-        "must go from CyaNewUkimsNumberPage to ???" in {
-
-          // TODO Needs to be updated according to navigation TGP-2700
-          navigator.nextPage(
-            CyaNewUkimsNumberPage,
-            NormalMode,
-            emptyUserAnswers
-          ) mustBe routes.IndexController.onPageLoad()
         }
       }
     }
@@ -3204,356 +2793,6 @@ class NavigatorSpec extends SpecBase with BeforeAndAfterEach {
             emptyUserAnswers
           ) mustBe routes.CyaSupplementaryUnitController
             .onPageLoad(testRecordId)
-        }
-      }
-
-      "in Create Profile Journey" - {
-        "must go from UkimsNumberPage to CyaCreateProfile" in {
-
-          navigator.nextPage(
-            UkimsNumberPage,
-            CheckMode,
-            emptyUserAnswers
-          ) mustBe routes.CyaCreateProfileController.onPageLoad()
-        }
-
-        "must go from HasNirmsPage" - {
-
-          "when answer is Yes" - {
-
-            "to NirmsNumberPage when NirmsNumberPage is empty" in {
-
-              val answers = UserAnswers(userAnswersId).set(HasNirmsPage, true).success.value
-              navigator.nextPage(HasNirmsPage, CheckMode, answers) mustBe routes.NirmsNumberController.onPageLoadCreate(
-                CheckMode
-              )
-            }
-
-            "to CyaCreateProfile when NirmsNumberPage is answered" in {
-
-              val answers =
-                UserAnswers(userAnswersId)
-                  .set(HasNirmsPage, true)
-                  .success
-                  .value
-                  .set(NirmsNumberPage, "1234")
-                  .success
-                  .value
-              navigator.nextPage(HasNirmsPage, CheckMode, answers) mustBe routes.CyaCreateProfileController.onPageLoad()
-            }
-          }
-          "to CyaCreateProfile when answer is No" in {
-
-            val answers = UserAnswers(userAnswersId).set(HasNirmsPage, false).success.value
-            navigator.nextPage(HasNirmsPage, CheckMode, answers) mustBe routes.CyaCreateProfileController.onPageLoad()
-          }
-
-          "to JourneyRecoveryPage when answer is not present" in {
-
-            navigator.nextPage(
-              HasNirmsPage,
-              CheckMode,
-              emptyUserAnswers
-            ) mustBe routes.JourneyRecoveryController
-              .onPageLoad()
-          }
-        }
-
-        "must go from NirmsNumberPage to CyaCreateProfile" in {
-
-          navigator.nextPage(
-            NirmsNumberPage,
-            CheckMode,
-            emptyUserAnswers
-          ) mustBe routes.CyaCreateProfileController.onPageLoad()
-        }
-
-        "must go from HasNiphlPage" - {
-
-          "when answer is Yes" - {
-
-            "to NiphlNumberPage when NiphlNumberPage is empty" in {
-
-              val answers = UserAnswers(userAnswersId).set(HasNiphlPage, true).success.value
-              navigator.nextPage(HasNiphlPage, CheckMode, answers) mustBe routes.NiphlNumberController.onPageLoadCreate(
-                CheckMode
-              )
-            }
-
-            "to CyaCreateProfile when NiphlNumberPage is answered" in {
-
-              val answers =
-                UserAnswers(userAnswersId)
-                  .set(HasNiphlPage, true)
-                  .success
-                  .value
-                  .set(NiphlNumberPage, "1234")
-                  .success
-                  .value
-              navigator.nextPage(HasNiphlPage, CheckMode, answers) mustBe routes.CyaCreateProfileController.onPageLoad()
-            }
-          }
-
-          "to CyaCreateProfile when answer is No" in {
-
-            val answers = UserAnswers(userAnswersId).set(HasNiphlPage, false).success.value
-            navigator.nextPage(HasNiphlPage, CheckMode, answers) mustBe routes.CyaCreateProfileController.onPageLoad()
-          }
-
-          "to JourneyRecoveryPage when answer is not present" in {
-
-            navigator.nextPage(
-              HasNiphlPage,
-              CheckMode,
-              emptyUserAnswers
-            ) mustBe routes.JourneyRecoveryController
-              .onPageLoad()
-          }
-        }
-
-        "must go from NiphlNumberPage to CyaCreateProfile" in {
-
-          navigator.nextPage(
-            NiphlNumberPage,
-            CheckMode,
-            emptyUserAnswers
-          ) mustBe routes.CyaCreateProfileController.onPageLoad()
-        }
-      }
-
-      "in Update Profile Journey" - {
-
-        "must go from UkimsNumberUpdatePage to CyaMaintainProfilePage" in {
-
-          navigator.nextPage(
-            UkimsNumberUpdatePage,
-            CheckMode,
-            emptyUserAnswers
-          ) mustBe routes.CyaMaintainProfileController.onPageLoadUkimsNumber()
-        }
-
-        "must go from RemoveNirmsPage" - {
-
-          "to CyaMaintainProfile when user answered No and NimrsNumberUpdate is defined" in {
-
-            val answers = UserAnswers(userAnswersId)
-              .set(RemoveNirmsPage, false)
-              .success
-              .value
-              .set(NirmsNumberUpdatePage, "some nirms")
-              .success
-              .value
-
-            navigator.nextPage(
-              RemoveNirmsPage,
-              CheckMode,
-              answers
-            ) mustBe routes.CyaMaintainProfileController.onPageLoadNirmsNumber()
-          }
-
-          "to CyaMaintainProfile when user answered No" in {
-
-            val answers = UserAnswers(userAnswersId).set(RemoveNirmsPage, false).success.value
-
-            navigator.nextPage(
-              RemoveNirmsPage,
-              CheckMode,
-              answers
-            ) mustBe routes.CyaMaintainProfileController.onPageLoadNirmsNumber()
-          }
-
-          "to Cya NIRMS registered when user answered yes" in {
-
-            val answers = UserAnswers(userAnswersId).set(RemoveNirmsPage, true).success.value
-
-            navigator.nextPage(
-              RemoveNirmsPage,
-              CheckMode,
-              answers
-            ) mustBe routes.CyaMaintainProfileController.onPageLoadNirms()
-          }
-
-          "must go from NirmsNumberUpdatePage to CyaMaintainProfile" in {
-
-            navigator.nextPage(
-              NirmsNumberUpdatePage,
-              CheckMode,
-              emptyUserAnswers
-            ) mustBe routes.CyaMaintainProfileController.onPageLoadNirmsNumber()
-          }
-        }
-
-        "must go from HasNirmsUpdatePage" - {
-
-          "to NirmsNumberUpdatePage when answer is Yes" in {
-
-            val answers = UserAnswers(userAnswersId).set(HasNirmsUpdatePage, true).success.value
-            navigator.nextPage(
-              HasNirmsUpdatePage,
-              CheckMode,
-              answers
-            ) mustBe routes.NirmsNumberController.onPageLoadUpdate(CheckMode)
-          }
-
-          "to RemoveNirmsPage when answer is No and Nirms number is associated to profile" in {
-
-            val answers = UserAnswers(userAnswersId)
-              .set(HasNirmsUpdatePage, false)
-              .success
-              .value
-              .set(
-                TraderProfileQuery,
-                TraderProfile("actorId", "ukims", Some("nirms"), Some("niphl"), eoriChanged = false)
-              )
-              .success
-              .value
-            navigator.nextPage(HasNirmsUpdatePage, CheckMode, answers) mustBe routes.RemoveNirmsController
-              .onPageLoad()
-          }
-
-          "to RemoveNirmsPage when answer is No and Nirms number is not associated to profile" in {
-
-            val answers = UserAnswers(userAnswersId)
-              .set(HasNirmsUpdatePage, false)
-              .success
-              .value
-              .set(TraderProfileQuery, TraderProfile("actorId", "ukims", None, Some("niphl"), eoriChanged = false))
-              .success
-              .value
-
-            navigator.nextPage(
-              HasNirmsUpdatePage,
-              CheckMode,
-              answers
-            ) mustBe routes.CyaMaintainProfileController.onPageLoadNirms()
-          }
-
-          "to JourneyRecoveryPage when answer is not present" in {
-            val continueUrl = RedirectUrl(routes.ProfileController.onPageLoad().url)
-
-            navigator.nextPage(
-              HasNirmsUpdatePage,
-              CheckMode,
-              emptyUserAnswers
-            ) mustBe routes.JourneyRecoveryController
-              .onPageLoad(Some(continueUrl))
-          }
-
-          "to JourneyRecoveryPage when TraderProfileQuery not present" in {
-            val answers = UserAnswers(userAnswersId)
-              .set(HasNirmsUpdatePage, false)
-              .success
-              .value
-
-            val continueUrl = RedirectUrl(routes.ProfileController.onPageLoad().url)
-
-            navigator.nextPage(
-              HasNirmsUpdatePage,
-              CheckMode,
-              answers
-            ) mustBe routes.JourneyRecoveryController
-              .onPageLoad(Some(continueUrl))
-          }
-        }
-
-        "must go from NirmsNumberUpdatePage to CyaMaintainProfile" in {
-
-          navigator.nextPage(
-            NirmsNumberUpdatePage,
-            CheckMode,
-            emptyUserAnswers
-          ) mustBe routes.CyaMaintainProfileController.onPageLoadNirmsNumber()
-        }
-
-        "must go from HasNiphlUpdatePage" - {
-
-          "to NiphlNumberUpdatePage when answer is Yes" in {
-
-            val answers = UserAnswers(userAnswersId).set(HasNiphlUpdatePage, true).success.value
-            navigator.nextPage(
-              HasNiphlUpdatePage,
-              CheckMode,
-              answers
-            ) mustBe routes.NiphlNumberController.onPageLoadUpdate(CheckMode)
-          }
-
-          "to RemoveNiphlPage when answer is No and Niphl number is associated to profile" in {
-
-            val answers = UserAnswers(userAnswersId)
-              .set(HasNiphlUpdatePage, false)
-              .success
-              .value
-              .set(
-                TraderProfileQuery,
-                TraderProfile("actorId", "ukims", Some("nirms"), Some("niphl"), eoriChanged = false)
-              )
-              .success
-              .value
-
-            navigator.nextPage(HasNiphlUpdatePage, CheckMode, answers) mustBe routes.RemoveNiphlController.onPageLoad()
-          }
-
-          "to RemoveNiphlPage when answer is No and Niphl number is not associated to profile" in {
-
-            val answers = UserAnswers(userAnswersId)
-              .set(HasNiphlUpdatePage, false)
-              .success
-              .value
-              .set(TraderProfileQuery, TraderProfile("actorId", "ukims", Some("nirms"), None, eoriChanged = false))
-              .success
-              .value
-
-            navigator.nextPage(
-              HasNiphlUpdatePage,
-              CheckMode,
-              answers
-            ) mustBe routes.CyaMaintainProfileController.onPageLoadNiphl()
-          }
-
-          "to JourneyRecoveryPage when answer is not present" in {
-            val continueUrl = RedirectUrl(routes.ProfileController.onPageLoad().url)
-
-            navigator.nextPage(
-              HasNiphlUpdatePage,
-              CheckMode,
-              emptyUserAnswers
-            ) mustBe routes.JourneyRecoveryController
-              .onPageLoad(Some(continueUrl))
-          }
-
-          "to JourneyRecoveryPage when TraderProfileQuery not present" in {
-            val answers = UserAnswers(userAnswersId)
-              .set(HasNiphlUpdatePage, false)
-              .success
-              .value
-
-            val continueUrl = RedirectUrl(routes.ProfileController.onPageLoad().url)
-
-            navigator.nextPage(
-              HasNiphlUpdatePage,
-              CheckMode,
-              answers
-            ) mustBe routes.JourneyRecoveryController
-              .onPageLoad(Some(continueUrl))
-          }
-        }
-
-        "must go from NiphlNumberUpdatePage to CyaMaintainProfile" in {
-
-          navigator.nextPage(
-            NiphlNumberUpdatePage,
-            CheckMode,
-            emptyUserAnswers
-          ) mustBe routes.CyaMaintainProfileController.onPageLoadNiphlNumber()
-        }
-
-        "must go from CyaMaintainProfilePage to CyaMaintainProfile" in {
-
-          navigator.nextPage(
-            CyaMaintainProfilePage,
-            CheckMode,
-            emptyUserAnswers
-          ) mustBe routes.ProfileController.onPageLoad()
         }
       }
 
@@ -5249,7 +4488,6 @@ class NavigatorSpec extends SpecBase with BeforeAndAfterEach {
         }
 
       }
-
     }
 
     ".journeyRecovery" - {
